Renaming ItemuTorrenting to ItemTorrentDownload
